feat(web): semantic task + assets auto-download; reuse visualizer palette
- Reuse the palette + pose skeleton from their original home in visualizer/
  (the native window viewer is already feature-gated, so the modules compile on
  wasm); delete the extra colors.rs. Single source, less code.
- Semantic segmentation: build_mask_overlay now also renders the semantic class
  map as a translucent RGBA overlay (50/50 blend, matching the native renderer).
- Auto-download: a bare ONNX name (yolo26/yolo11/yolov8, n/s/m/l/x + task
  suffixes) resolves to the Ultralytics assets release; URLs/paths pass through.
- LICENSE: generate web/LICENSE from the repo-root LICENSE at pack time
  (prepack) instead of committing a 661-line duplicate.
- Document auto-download + semantic in the npm README.
