+++ web/package.json
@@ -29,7 +29,8 @@
     "build:wasm": "wasm-pack build ../crates/web --release --target web --out-dir \"$PWD/pkg\" --out-name ultralytics_inference_web && rm -f \"$PWD/pkg/.gitignore\"",
     "build:ts": "tsc",
     "build": "npm run build:wasm && npm run build:ts",
-    "clean": "rm -rf dist pkg"
+    "prepack": "cp ../LICENSE ./LICENSE",
+    "clean": "rm -rf dist pkg LICENSE"
   },
   "keywords": [
     "yolo",
